+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/DirectionUploadOrderAboutToStartHandler.java
@@ -1,12 +1,10 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.handler.processorders;
+package uk.gov.hmcts.reform.finrem.caseorchestration.handler;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.finrem.caseorchestration.ccd.callback.CallbackType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.controllers.GenericAboutToStartOrSubmitCallbackResponse;
-import uk.gov.hmcts.reform.finrem.caseorchestration.handler.FinremCallbackHandler;
-import uk.gov.hmcts.reform.finrem.caseorchestration.handler.FinremCallbackRequest;
 import uk.gov.hmcts.reform.finrem.caseorchestration.mapper.FinremCaseDetailsMapper;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType;
@@ -20,24 +18,24 @@ import java.util.List;
 
 import static java.util.Optional.ofNullable;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.ccd.callback.CallbackType.ABOUT_TO_START;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType.PROCESS_ORDER;
+import static uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType.DIRECTION_UPLOAD_ORDER;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType.CONTESTED;
 
 @Slf4j
 @Service
-public class ProcessOrdersAboutToStartHandler extends FinremCallbackHandler {
+public class DirectionUploadOrderAboutToStartHandler extends FinremCallbackHandler {
 
     private final ProcessOrderService processOrderService;
 
-    public ProcessOrdersAboutToStartHandler(FinremCaseDetailsMapper finremCaseDetailsMapper,
-                                            ProcessOrderService processOrderService) {
+    public DirectionUploadOrderAboutToStartHandler(FinremCaseDetailsMapper finremCaseDetailsMapper,
+                                                   ProcessOrderService processOrderService) {
         super(finremCaseDetailsMapper);
         this.processOrderService = processOrderService;
     }
 
     @Override
     public boolean canHandle(CallbackType callbackType, CaseType caseType, EventType eventType) {
-        return ABOUT_TO_START.equals(callbackType) && CONTESTED.equals(caseType) && PROCESS_ORDER.equals(eventType);
+        return ABOUT_TO_START.equals(callbackType) && CONTESTED.equals(caseType) && DIRECTION_UPLOAD_ORDER.equals(eventType);
     }
 
     @Override
