+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/DirectionUploadOrderMidHandler.java
@@ -7,25 +7,33 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.controllers.GenericAboutToSt
 import uk.gov.hmcts.reform.finrem.caseorchestration.mapper.FinremCaseDetailsMapper;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.DirectionDetail;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.DirectionDetailCollection;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.DirectionOrderCollection;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.DocumentCollection;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseData;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseDetails;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.BulkPrintDocumentService;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.processorder.ProcessOrderService;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import static java.util.Optional.ofNullable;
 
 @Slf4j
 @Service
 public class DirectionUploadOrderMidHandler extends FinremCallbackHandler {
 
-    private final BulkPrintDocumentService service;
+    private final BulkPrintDocumentService bulkPrintDocumentService;
+
+    private final ProcessOrderService processOrderService;
 
     public DirectionUploadOrderMidHandler(FinremCaseDetailsMapper finremCaseDetailsMapper,
-                                          BulkPrintDocumentService service) {
+                                          BulkPrintDocumentService bulkPrintDocumentService, ProcessOrderService processOrderService) {
         super(finremCaseDetailsMapper);
-        this.service = service;
+        this.bulkPrintDocumentService = bulkPrintDocumentService;
+        this.processOrderService = processOrderService;
     }
 
     @Override
@@ -48,6 +56,11 @@ public class DirectionUploadOrderMidHandler extends FinremCallbackHandler {
         FinremCaseDetails caseDetailsBefore = callbackRequest.getCaseDetailsBefore();
         FinremCaseData caseDataBefore = caseDetailsBefore.getData();
 
+        if (processOrderService.areAllLegacyApprovedOrdersRemoved(caseDataBefore, caseData)) {
+            return GenericAboutToStartOrSubmitCallbackResponse.<FinremCaseData>builder()
+                .data(caseData).errors(List.of("Upload Approved Order is required.")).build();
+        }
+
         List<DirectionOrderCollection> uploadHearingOrders = caseData.getUploadHearingOrder();
         if (uploadHearingOrders != null) {
             List<DirectionOrderCollection> uploadHearingOrdersBefore = caseDataBefore.getUploadHearingOrder();
@@ -55,7 +68,7 @@ public class DirectionUploadOrderMidHandler extends FinremCallbackHandler {
                 uploadHearingOrders.removeAll(uploadHearingOrdersBefore);
             }
             uploadHearingOrders.forEach(doc ->
-                service.validateEncryptionOnUploadedDocument(doc.getValue().getUploadDraftDocument(),
+                bulkPrintDocumentService.validateEncryptionOnUploadedDocument(doc.getValue().getUploadDraftDocument(),
                     caseId, errors, userAuthorisation)
             );
         }
@@ -66,9 +79,33 @@ public class DirectionUploadOrderMidHandler extends FinremCallbackHandler {
                 hearingOrderOtherDocuments.removeAll(hearingOrderOtherDocumentsBefore);
             }
             hearingOrderOtherDocuments.forEach(doc ->
-                service.validateEncryptionOnUploadedDocument(doc.getValue(),
+                bulkPrintDocumentService.validateEncryptionOnUploadedDocument(doc.getValue(),
                     caseId, errors, userAuthorisation)
             );
+        }
+
+        processOrderService.populateUnprocessedApprovedDocuments(caseDataBefore);
+        if (!processOrderService.areAllNewOrdersPdfFiles(caseDataBefore, caseData)) {
+            return GenericAboutToStartOrSubmitCallbackResponse.<FinremCaseData>builder()
+                .data(caseData).errors(List.of("You must upload a PDF file for new documents.")).build();
+        }
+        // Validate the modifying legacy approved orders
+        if (!processOrderService.areAllLegacyApprovedOrdersPdf(caseData)) {
+            return GenericAboutToStartOrSubmitCallbackResponse.<FinremCaseData>builder()
+                .data(caseData).errors(List.of("You must upload a PDF file for modifying legacy approved documents.")).build();
+        }
+        // Validate the modifying unprocessed approved orders are word documents
+        if (!processOrderService.areAllModifyingUnprocessedOrdersWordDocuments(caseData)) {
+            return GenericAboutToStartOrSubmitCallbackResponse.<FinremCaseData>builder()
+                .data(caseData).errors(List.of("You must upload a Microsoft Word file for modifying an unprocessed approved documents."))
+                .build();
+        }
+
+        // Create an empty entry if it is empty to save a click on add new button
+        if (ofNullable(caseData.getDirectionDetailsCollection()).orElse(List.of()).isEmpty()) {
+            caseData.setDirectionDetailsCollection(List.of(
+                DirectionDetailCollection.builder().value(DirectionDetail.builder().build()).build()
+            ));
         }
 
         return GenericAboutToStartOrSubmitCallbackResponse.<FinremCaseData>builder()
